refactor auth plugin for testability add config for JWT decode failure behavior apply RustRover lints move authentication errors into a module move jwks utils into a module
