+++ apollo-router/src/plugins/authentication/jwks.rs
@@ -2,7 +2,7 @@ use std::collections::HashMap;
 use std::collections::HashSet;
 use std::mem;
 use std::sync::Arc;
-use std::time::Duration;
+use std::time::{Duration, SystemTime, UNIX_EPOCH};
 
 use futures::future::Either;
 use futures::future::join_all;
@@ -11,8 +11,8 @@ use futures::pin_mut;
 use futures::stream::repeat;
 use futures::stream::select_all;
 use http::header::ACCEPT;
-use jsonwebtoken::Algorithm;
-use jsonwebtoken::jwk::Jwk;
+use jsonwebtoken::{decode, Algorithm, DecodingKey, TokenData, Validation};
+use jsonwebtoken::jwk::{AlgorithmParameters, EllipticCurve, Jwk, KeyAlgorithm, KeyOperations, PublicKeyUse};
 use jsonwebtoken::jwk::JwkSet;
 use mime::APPLICATION_JSON;
 use parking_lot::RwLock;
@@ -22,8 +22,10 @@ use tokio::sync::oneshot;
 use tower::BoxError;
 use tracing_futures::Instrument;
 use url::Url;
-
-use super::CLIENT;
+use http::{HeaderMap, StatusCode};
+use crate::Context;
+use crate::plugins::authentication::error::AuthenticationError;
+use super::{Source, APOLLO_AUTHENTICATION_JWT_CLAIMS, CLIENT};
 use super::DEFAULT_AUTHENTICATION_NETWORK_TIMEOUT;
 use super::Header;
 
@@ -137,7 +139,7 @@ async fn poll(
 }
 
 // This function is expected to return an Optional value, but we'd like to let
-// users know the various failure conditions. Hence the various clumsy map_err()
+// users know the various failure conditions. Hence, the various clumsy map_err()
 // scattered through the processing.
 pub(super) async fn get_jwks(url: Url, headers: Vec<Header>) -> Option<JwkSet> {
     let data = if url.scheme() == "file" {
@@ -254,5 +256,424 @@ impl Iterator for Iter<'_> {
                 }
             }
         }
+    }
+}
+
+#[derive(Debug, Default)]
+pub(super) struct JWTCriteria {
+    pub(super) alg: Algorithm,
+    pub(super) kid: Option<String>,
+}
+
+/// Search the list of JWKS to find a key we can use to decode a JWT.
+///
+/// The search criteria allow us to match a variety of keys depending on which criteria are provided
+/// by the JWT header. The only mandatory parameter is "alg".
+/// Note: "none" is not implemented by jsonwebtoken, so it can't be part of the [`Algorithm`] enum.
+pub(super) fn search_jwks(
+    jwks_manager: &JwksManager,
+    criteria: &JWTCriteria,
+) -> Option<Vec<(Option<String>, Jwk)>> {
+    const HIGHEST_SCORE: usize = 2;
+    let mut candidates = vec![];
+    let mut found_highest_score = false;
+    for JwkSetInfo {
+        jwks,
+        issuer,
+        algorithms,
+    } in jwks_manager.iter_jwks()
+    {
+        // filter accepted algorithms
+        if let Some(algs) = algorithms {
+            if !algs.contains(&criteria.alg) {
+                continue;
+            }
+        }
+
+        // Try to figure out if our jwks contains a candidate key (i.e.: a key which matches our
+        // criteria)
+        for mut key in jwks.keys.into_iter().filter(|key| {
+            // We are only interested in keys which are used for signature verification
+            match (&key.common.public_key_use, &key.common.key_operations) {
+                // "use" https://datatracker.ietf.org/doc/html/rfc7517#section-4.2 and
+                // "key_ops" https://datatracker.ietf.org/doc/html/rfc7517#section-4.3 are both optional
+                (None, None) => true,
+                (None, Some(purpose)) => purpose.contains(&KeyOperations::Verify),
+                (Some(key_use), None) => key_use == &PublicKeyUse::Signature,
+                // The "use" and "key_ops" JWK members SHOULD NOT be used together;
+                // however, if both are used, the information they convey MUST be
+                // consistent
+                (Some(key_use), Some(purpose)) => {
+                    key_use == &PublicKeyUse::Signature && purpose.contains(&KeyOperations::Verify)
+                }
+            }
+        }) {
+            let mut key_score = 0;
+
+            // Let's see if we have a specified kid and if they match
+            if criteria.kid.is_some() && key.common.key_id == criteria.kid {
+                key_score += 1;
+            }
+
+            // Furthermore, we would like our algorithms to match, or at least the kty
+            // If we have an algorithm that matches, boost the score
+            match key.common.key_algorithm {
+                Some(algorithm) => {
+                    if convert_key_algorithm(algorithm) != Some(criteria.alg) {
+                        continue;
+                    }
+                    key_score += 1;
+                }
+                // If a key doesn't have an algorithm, then we match the "alg" specified in the
+                // search criteria against all of the algorithms that we support.  If the
+                // key.algorithm parameters match the type of parameters for the "family" of the
+                // criteria "alg", then we update the key to use the value of "alg" provided in
+                // the search criteria.
+                // If not, then this is not a usable key for this JWT
+                // Note: Matching algorithm parameters may seem unusual, but the appropriate
+                // algorithm details are not structured for easy consumption in jsonwebtoken and
+                // this is the simplest way to determine algorithm family.
+                None => match (criteria.alg, &key.algorithm) {
+                    (
+                        Algorithm::HS256 | Algorithm::HS384 | Algorithm::HS512,
+                        AlgorithmParameters::OctetKey(_),
+                    ) => {
+                        key.common.key_algorithm = Some(convert_algorithm(criteria.alg));
+                    }
+                    (
+                        Algorithm::RS256
+                        | Algorithm::RS384
+                        | Algorithm::RS512
+                        | Algorithm::PS256
+                        | Algorithm::PS384
+                        | Algorithm::PS512,
+                        AlgorithmParameters::RSA(_),
+                    ) => {
+                        key.common.key_algorithm = Some(convert_algorithm(criteria.alg));
+                    }
+                    (Algorithm::ES256, AlgorithmParameters::EllipticCurve(params)) => {
+                        if params.curve == EllipticCurve::P256 {
+                            key.common.key_algorithm = Some(convert_algorithm(criteria.alg));
+                        }
+                    }
+                    (Algorithm::ES384, AlgorithmParameters::EllipticCurve(params)) => {
+                        if params.curve == EllipticCurve::P384 {
+                            key.common.key_algorithm = Some(convert_algorithm(criteria.alg));
+                        }
+                    }
+                    (Algorithm::EdDSA, AlgorithmParameters::EllipticCurve(params)) => {
+                        if params.curve == EllipticCurve::Ed25519 {
+                            key.common.key_algorithm = Some(convert_algorithm(criteria.alg));
+                        }
+                    }
+                    _ => {
+                        // We'll ignore combinations we don't recognise
+                        continue;
+                    }
+                },
+            };
+
+            // If we get here we have a key that:
+            //  - may be used for signature verification
+            //  - has a matching algorithm, or if JWT has no algorithm, a matching key type
+            // It may have a matching kid if the JWT has a kid and it matches the key kid
+            //
+            // Multiple keys may meet the matching criteria, but they have a score. They get 1
+            // point for having an explicitly matching algorithm and 1 point for an explicitly
+            // matching kid. We will sort our candidates and pick the key with the highest score.
+
+            // If we find a key with a HIGHEST_SCORE, we will filter the list to only keep those
+            // with that score
+            if key_score == HIGHEST_SCORE {
+                found_highest_score = true;
+            }
+
+            candidates.push((key_score, (issuer.clone(), key)));
+        }
+    }
+
+    tracing::debug!(
+        "jwk candidates: {:?}",
+        candidates
+            .iter()
+            .map(|(score, (_, candidate))| (
+                score,
+                &candidate.common.key_id,
+                candidate.common.key_algorithm
+            ))
+            .collect::<Vec<(&usize, &Option<String>, Option<KeyAlgorithm>)>>()
+    );
+
+    if candidates.is_empty() {
+        None
+    } else {
+        // Only sort if we need to
+        if candidates.len() > 1 {
+            candidates.sort_by(|a, b| a.0.cmp(&b.0));
+        }
+
+        if found_highest_score {
+            Some(
+                candidates
+                    .into_iter()
+                    .filter_map(|(score, candidate)| {
+                        if score == HIGHEST_SCORE {
+                            Some(candidate)
+                        } else {
+                            None
+                        }
+                    })
+                    .collect(),
+            )
+        } else {
+            Some(
+                candidates
+                    .into_iter()
+                    .map(|(_score, candidate)| candidate)
+                    .collect(),
+            )
+        }
+    }
+}
+
+pub(super) fn extract_jwt<'a, 'b: 'a>(
+    source: &'a Source,
+    ignore_other_prefixes: bool,
+    headers: &'b HeaderMap,
+) -> Option<Result<&'b str, AuthenticationError<'a>>> {
+    match source {
+        Source::Header { name, value_prefix } => {
+            // The http_request is stored in a `Router::Request` context.
+            // We are going to check the headers for the presence of the configured header
+            let jwt_value_result = match headers.get(name) {
+                Some(value) => value.to_str(),
+                None => return None,
+            };
+
+            // If we find the header, but can't convert it to a string, let the client know
+            let jwt_value_untrimmed = match jwt_value_result {
+                Ok(value) => value,
+                Err(_not_a_string_error) => {
+                    return Some(Err(AuthenticationError::CannotConvertToString));
+                }
+            };
+
+            // Let's trim out leading and trailing whitespace to be accommodating
+            let jwt_value = jwt_value_untrimmed.trim();
+
+            // Make sure the format of our message matches our expectations
+            // Technically, the spec is case sensitive, but let's accept
+            // case variations
+            //
+            let prefix_len = value_prefix.len();
+            if jwt_value.len() < prefix_len
+                || !&jwt_value[..prefix_len].eq_ignore_ascii_case(value_prefix)
+            {
+                return if ignore_other_prefixes {
+                    None
+                } else {
+                    Some(Err(AuthenticationError::InvalidPrefix(
+                        jwt_value_untrimmed,
+                        value_prefix,
+                    )))
+                };
+            }
+            // If there's no header prefix, we need to avoid splitting the header
+            let jwt = if value_prefix.is_empty() {
+                // check for whitespace- we've already trimmed, so this means the request has a prefix that shouldn't exist
+                if jwt_value.contains(' ') {
+                    return Some(Err(AuthenticationError::InvalidPrefix(
+                        jwt_value_untrimmed,
+                        value_prefix,
+                    )));
+                }
+
+                // we can simply assign the jwt to the jwt_value; we'll validate down below
+                jwt_value
+            } else {
+                // Otherwise, we need to split our string in (at most 2) sections.
+                let jwt_parts: Vec<&str> = jwt_value.splitn(2, ' ').collect();
+                if jwt_parts.len() != 2 {
+                    return Some(Err(AuthenticationError::MissingJWT(jwt_value)));
+                }
+
+                // We have our jwt
+                jwt_parts[1]
+            };
+            Some(Ok(jwt))
+        }
+        Source::Cookie { name } => {
+            for header in headers.get_all("cookie") {
+                let value = match header.to_str() {
+                    Ok(value) => value,
+                    Err(_not_a_string_error) => {
+                        return Some(Err(AuthenticationError::CannotConvertToString));
+                    }
+                };
+                for cookie in cookie::Cookie::split_parse(value) {
+                    match cookie {
+                        Err(_) => continue,
+                        Ok(cookie) => {
+                            if cookie.name() == name {
+                                if let Some(value) = cookie.value_raw() {
+                                    return Some(Ok(value));
+                                }
+                            }
+                        }
+                    }
+                }
+            }
+
+            None
+        }
+    }
+}
+
+pub(super) fn decode_jwt(
+    jwt: &str,
+    keys: Vec<(Option<String>, Jwk)>,
+    criteria: JWTCriteria,
+) -> Result<(Option<String>, TokenData<serde_json::Value>), (AuthenticationError, StatusCode)> {
+    let mut error = None;
+    for (issuer, jwk) in keys.into_iter() {
+        let decoding_key = match DecodingKey::from_jwk(&jwk) {
+            Ok(k) => k,
+            Err(e) => {
+                error = Some((
+                    AuthenticationError::CannotCreateDecodingKey(e),
+                    StatusCode::INTERNAL_SERVER_ERROR,
+                ));
+                continue;
+            }
+        };
+
+        let key_algorithm = match jwk.common.key_algorithm {
+            Some(a) => a,
+            None => {
+                error = Some((
+                    AuthenticationError::JWKHasNoAlgorithm,
+                    StatusCode::INTERNAL_SERVER_ERROR,
+                ));
+                continue;
+            }
+        };
+
+        let algorithm = match convert_key_algorithm(key_algorithm) {
+            Some(a) => a,
+            None => {
+                error = Some((
+                    AuthenticationError::UnsupportedKeyAlgorithm(key_algorithm),
+                    StatusCode::INTERNAL_SERVER_ERROR,
+                ));
+                continue;
+            }
+        };
+
+        let mut validation = Validation::new(algorithm);
+        validation.validate_nbf = true;
+        // if set to true, it will reject tokens containing an `aud` claim if the validation does not specify an audience
+        // we don't validate audience yet, so this is deactivated
+        validation.validate_aud = false;
+
+        match decode::<serde_json::Value>(jwt, &decoding_key, &validation) {
+            Ok(v) => return Ok((issuer, v)),
+            Err(e) => {
+                tracing::trace!("JWT decoding failed with error `{e}`");
+                error = Some((
+                    AuthenticationError::CannotDecodeJWT(e),
+                    StatusCode::UNAUTHORIZED,
+                ));
+            }
+        };
+    }
+
+    match error {
+        Some(e) => Err(e),
+        None => {
+            // We can't find a key to process this JWT.
+            if criteria.kid.is_some() {
+                Err((
+                    AuthenticationError::CannotFindKID(criteria.kid),
+                    StatusCode::UNAUTHORIZED,
+                ))
+            } else {
+                Err((
+                    AuthenticationError::CannotFindSuitableKey(criteria.alg, criteria.kid),
+                    StatusCode::UNAUTHORIZED,
+                ))
+            }
+        }
+    }
+}
+
+pub(crate) fn jwt_expires_in(context: &Context) -> Duration {
+    let claims = context
+        .get(APOLLO_AUTHENTICATION_JWT_CLAIMS)
+        .map_err(|err| tracing::error!("could not read JWT claims: {err}"))
+        .ok()
+        .flatten();
+    let ts_opt = claims.as_ref().and_then(|x: &serde_json::Value| {
+        if !x.is_object() {
+            tracing::error!("JWT claims should be an object");
+            return None;
+        }
+        let claims = x.as_object().expect("claims should be an object");
+        let exp = claims.get("exp")?;
+        if !exp.is_number() {
+            tracing::error!("JWT 'exp' (expiry) claim should be a number");
+            return None;
+        }
+        exp.as_i64()
+    });
+    match ts_opt {
+        Some(ts) => {
+            let now = SystemTime::now()
+                .duration_since(UNIX_EPOCH)
+                .expect("we should not run before EPOCH")
+                .as_secs() as i64;
+            if now < ts {
+                Duration::from_secs((ts - now) as u64)
+            } else {
+                Duration::ZERO
+            }
+        }
+        None => Duration::MAX,
+    }
+}
+
+// Apparently the `jsonwebtoken` crate now has 2 different enums for algorithms
+pub(crate) fn convert_key_algorithm(algorithm: KeyAlgorithm) -> Option<Algorithm> {
+    Some(match algorithm {
+        KeyAlgorithm::HS256 => Algorithm::HS256,
+        KeyAlgorithm::HS384 => Algorithm::HS384,
+        KeyAlgorithm::HS512 => Algorithm::HS512,
+        KeyAlgorithm::ES256 => Algorithm::ES256,
+        KeyAlgorithm::ES384 => Algorithm::ES384,
+        KeyAlgorithm::RS256 => Algorithm::RS256,
+        KeyAlgorithm::RS384 => Algorithm::RS384,
+        KeyAlgorithm::RS512 => Algorithm::RS512,
+        KeyAlgorithm::PS256 => Algorithm::PS256,
+        KeyAlgorithm::PS384 => Algorithm::PS384,
+        KeyAlgorithm::PS512 => Algorithm::PS512,
+        KeyAlgorithm::EdDSA => Algorithm::EdDSA,
+        // We don't use these encryption algorithms
+        KeyAlgorithm::RSA1_5 | KeyAlgorithm::RSA_OAEP | KeyAlgorithm::RSA_OAEP_256 => return None,
+    })
+}
+
+fn convert_algorithm(algorithm: Algorithm) -> KeyAlgorithm {
+    match algorithm {
+        Algorithm::HS256 => KeyAlgorithm::HS256,
+        Algorithm::HS384 => KeyAlgorithm::HS384,
+        Algorithm::HS512 => KeyAlgorithm::HS512,
+        Algorithm::ES256 => KeyAlgorithm::ES256,
+        Algorithm::ES384 => KeyAlgorithm::ES384,
+        Algorithm::RS256 => KeyAlgorithm::RS256,
+        Algorithm::RS384 => KeyAlgorithm::RS384,
+        Algorithm::RS512 => KeyAlgorithm::RS512,
+        Algorithm::PS256 => KeyAlgorithm::PS256,
+        Algorithm::PS384 => KeyAlgorithm::PS384,
+        Algorithm::PS512 => KeyAlgorithm::PS512,
+        Algorithm::EdDSA => KeyAlgorithm::EdDSA,
     }
 }
