+++ apollo-router/src/plugins/authentication/tests.rs
@@ -9,11 +9,12 @@ use axum::handler::HandlerWithoutStateExt;
 use base64::Engine as _;
 use base64::prelude::BASE64_URL_SAFE_NO_PAD;
 use http::header::CONTENT_TYPE;
+use http::HeaderMap;
 use insta::assert_yaml_snapshot;
 use jsonwebtoken::EncodingKey;
 use jsonwebtoken::encode;
 use jsonwebtoken::get_current_timestamp;
-use jsonwebtoken::jwk::CommonParameters;
+use jsonwebtoken::jwk::{AlgorithmParameters, CommonParameters, EllipticCurve, Jwk, KeyAlgorithm, KeyOperations, PublicKeyUse};
 use jsonwebtoken::jwk::EllipticCurveKeyParameters;
 use jsonwebtoken::jwk::EllipticCurveKeyType;
 use jsonwebtoken::jwk::JwkSet;
@@ -29,7 +30,7 @@ use super::Header;
 use super::*;
 use crate::assert_snapshot_subscriber;
 use crate::plugin::test;
-use crate::plugins::authentication::jwks::parse_jwks;
+use crate::plugins::authentication::jwks::{parse_jwks, search_jwks, JWTCriteria};
 use crate::services::router;
 use crate::services::router::body::RouterBody;
 use crate::services::supergraph;
